# server/internal/docs/swagger.yaml
basePath: /
definitions:
  github_com_hybrid-app_server_internal_domainutil.DomainInput:
    properties:
      enabled:
        type: boolean
      position:
        type: integer
      url:
        type: string
    type: object
  github_com_hybrid-app_server_internal_httpx.Envelope:
    properties:
      code:
        description: 0 = 成功，非 0 = 业务错误码（沿用 HTTP 状态码语义）
        type: integer
      data:
        description: 业务数据
      message:
        description: 人类可读信息
        type: string
    type: object
...
  github_com_hybrid-app_server_internal_service.CreateChannelInput:
    properties:
      adjustAppToken:
        type: string
      adjustEvents:
        additionalProperties:
          type: string
        type: object
      appName:
        type: string
      brandCode:
        type: string
      flavorName:
        type: string
      palCode:
        type: string
      remark:
        type: string
      storeId:
        type: integer
    required:
    - appName
    - brandCode
    - flavorName
    - palCode
    type: object
...
  github_com_hybrid-app_server_internal_service.UpdateChannelInput:
    properties:
      adjustAppToken:
        type: string
      adjustEvents:
        additionalProperties:
          type: string
        type: object
      appName:
        type: string
      flavorName:
        type: string
      palCode:
        type: string
      remark:
        type: string
      status:
        type: string
    type: object
...
  internal_handler.loginReq:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  internal_handler.meResp:
    properties:
      perms:
        items:
          type: string
        type: array
      role:
        $ref: '#/definitions/internal_handler.roleSummary'
      scope:
        $ref: '#/definitions/internal_handler.scopeSummary'
      username:
        type: string
    type: object
  internal_handler.refreshReq:
    properties:
      refreshToken:
        type: string
    type: object
...
  internal_handler.roleSummary:
    properties:
      builtin:
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
...
  internal_handler.scopeSummary:
    properties:
      allBrands:
        type: boolean
      allChannels:
        type: boolean
      brands:
        items:
          type: string
        type: array
      channelIds:
        items:
          type: integer
        type: array
    type: object
...
  internal_handler.setChannelDomainsReq:
    properties:
      domains:
        items:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_domainutil.DomainInput'
        type: array
      inheritBrand:
        type: boolean
    type: object
  internal_handler.setDomainsReq:
    properties:
      domains:
        items:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_domainutil.DomainInput'
        type: array
    type: object
...
  internal_handler.tokenResp:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      perms:
        description: 超级管理员返回完整 catalog code 列表
        items:
          type: string
        type: array
      refreshToken:
        type: string
      role:
        $ref: '#/definitions/internal_handler.roleSummary'
      scope:
        $ref: '#/definitions/internal_handler.scopeSummary'
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: hybrid-app 多渠道打包后台：渠道 CRUD / 图标管线 / 域名配置 / 运行时下发。
  title: 渠道中台 API
  version: "1.0"
paths:
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.tokenResp'
              type: object
      summary: 账号密码登录
      tags:
      - auth
  /api/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.meResp'
              type: object
      security:
      - BearerAuth: []
      summary: 当前登录用户的角色与权限(前端启动时拉新，角色被改后无需重登即可生效)
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.refreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_handler.tokenResp'
              type: object
      summary: 刷新 access token
      tags:
      - auth
  /api/brands:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 大渠道列表（含渠道计数、accent 色、默认域名；按调用者数据范围过滤）
      tags:
      - brands
  /api/brands/{code}/domains:
    get:
      parameters:
      - description: 品牌 code (ap/bp/gp)
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 品牌默认域名清单
      tags:
      - brands
    put:
      consumes:
      - application/json
      parameters:
      - description: 品牌 code
        in: path
        name: code
        required: true
        type: string
      - description: 域名清单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setDomainsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 更新品牌默认域名（主+最多3备用，强校验）
      tags:
      - brands
...
  /api/channels:
    get:
      parameters:
      - description: 品牌 code
        in: query
        name: brand
        type: string
      - description: 状态 enabled/disabled/archived
        in: query
        name: status
        type: string
      - description: 关键词（flavor/包名/应用名/palcode）
        in: query
        name: q
        type: string
      - description: 页码，默认 1
        in: query
        name: page
        type: integer
      - description: 每页，默认 50
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 渠道列表（分页/搜索/筛选；按调用者数据范围过滤）
      tags:
      - channels
    post:
      consumes:
      - application/json
      parameters:
      - description: 渠道字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_service.CreateChannelInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 新增渠道（applicationId 派生自品牌包前缀+flavor；唯一性校验 applicationId 与 (brand,flavor)）
      tags:
      - channels
  /api/channels/{id}:
    delete:
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 软删除渠道（置 archived；越界返回 404）
      tags:
      - channels
    get:
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 渠道详情（越界返回 404，不泄露存在性）
      tags:
      - channels
    put:
      consumes:
      - application/json
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 可选更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_service.UpdateChannelInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 修改渠道（越界返回 404）
      tags:
      - channels
  /api/channels/{id}/domains:
    put:
      consumes:
      - application/json
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      - description: inheritBrand=true 表示切回继承
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setChannelDomainsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 设置渠道域名覆盖 / 切回继承品牌
      tags:
      - channels
...
  /api/channels/{id}/latest-apk:
    get:
      parameters:
      - description: 渠道 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 渠道最新包：按该 flavor 取最近一次成功构建的 APK（ADR-0008 卡片「下载最新包」；越界返回 404）
      tags:
      - channels
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"